+++ BackEnd/ProjetoInt.WebApi/ProjetoInt.WebApi/Domains/FotoAluno.cs
@@ -8,7 +8,7 @@ namespace ProjetoInt.WebApi.Domains
     public partial class FotoAluno
     {
         public int IdFotoAluno { get; set; }
-        public short IdAluno { get; set; }
+        public int IdAluno { get; set; }
         public byte[] Binario { get; set; }
         public string MimeType { get; set; }
         public string NomeArquivo { get; set; }
